fixed menu issue
