+++ client/Food2Go/App.js
@@ -12,7 +12,6 @@ import Dish from "./screens/Dish";
 import Profile from "./screens/Profile";
 import Register from "./screens/Register";
 import ShopCart from "./screens/ShopCart";
-import NewMenu from "./screens/NewMenu";
 
 
 
@@ -64,7 +63,7 @@ export default function App() {
               </Stack.Screen>
               <Stack.Screen name="ListView"  component={ListView} />
               <Stack.Screen name="MapScreen"  component={MapScreen} />
-              <Stack.Screen name="Menu"  component={NewMenu} />
+              <Stack.Screen name="Menu"  component={Menu} />
               <Stack.Screen name="Dish" component={Dish} />
               <Stack.Screen name="Profile" component={Profile} />
               <Stack.Screen name="ShopCart" component={ShopCart} />
